Overhaul nonce storage, eliminate memory leak
1.  Use v4 (random) UUIDs instead of binascii/urandom gymnastics.

    It works similarly under the hood, but has a nicer interface.

2.  Stop keying nonces on email addresses.

    Users may have multiple logins in flight at once, such as both a
    laptop and a phone. The previous design prevented that.

    I'm also not sure we gain any additional benefit by explicitly tying
    the nonce to an email address, so we might as well keep it simple.

3.  Make nonces expire after 30 minutes, periodically filter expired.

    This mitigates a memory leak in the previous design, whereby unused
    nonces were never removed from the dict. It would still be possible
    to attack this by initiating, but not completing, many logins.

I'd strongly prefer to use Redis or a similar key/value store with
built-in record expiration so that this isn't our problem. That also
more closely match what developers must do in production, while the
current design is completely inapplicable beyond local development.

I'm inclined to drop nonces altogether from this implementation, but it
is still useful for testing the broker, so... in it stays.