Fix b64dec to not pad when len(s) % 4 == 4
The previous function would unnecessarily pad with '====' when the
string length was already a multiple of 4. Adding a further `% 4`
resolves this issue.

Not a problem for Python, since it allows superfluous padding, but
other languages are often more strict.

Previous:

    >>> pad = lambda s: s + '=' * (4 - len(s) % 4)
    >>> [pad(s) for s in ('', 'a', 'ab', 'abc', 'abcd')]
    ['====', 'a===', 'ab==', 'abc=', 'abcd====']

New:

    >>> pad = lambda s: s + '=' * ((4 - len(s) % 4) % 4)
    >>> [pad(s) for s in ('', 'a', 'ab', 'abc', 'abcd')]
    ['', 'a===', 'ab==', 'abc=', 'abcd']

We can also skip the call to s.encode('ascii'), since the Python
standard library automatically converts strings to bytestrings:
https://hg.python.org/cpython/file/3.5/Lib/base64.py#l132