Switch to .ini based config, support using envvars
Ini files allow comments, which make it easier to document the settings.

This also makes it possible to override settings using environment
variables, a prerequisite for hosting on PaaS providers like Heroku.
More info and rationale at https://11factor.net/config

Lastly, this separates the notion of the RP's URL and its host/port:
The webserver needs to know what host/port to listen on, while the
Broker needs to know what publicly visible URL to redirect users to.

For example, the demo may be publicly accessible at example.com, but
actually listening on 127.0.0.1:8000 behind an Apache or nginx server.